lwm2m_client: update to be compatible with current Zephyr version
Remove undefined CONFIG_POSIX_CLOCK and duplicate
CONFIG_LWM2M_RW_SENML_CBOR_SUPPORT from prj.conf. Update network
management event handlers to use uint64_t for event type. Fix
firmware_update.c to use Zephyr POSIX headers and add missing includes.

+++ simulation/lwm2m_client/src/lwm2m-client.c
@@ -343,10 +343,10 @@ static void on_net_event_l4_connected(void)
 }
 
 static void l4_event_handler(struct net_mgmt_event_callback *cb,
-			     uint32_t event,
+			     uint64_t event,
 			     struct net_if *iface)
 {
-	LOG_DBG("Event %d", event);
+	LOG_DBG("Event %llu", event);
 
 	switch (event) {
 	case NET_EVENT_L4_CONNECTED:
@@ -363,10 +363,10 @@ static void l4_event_handler(struct net_mgmt_event_callback *cb,
 }
 
 static void connectivity_event_handler(struct net_mgmt_event_callback *cb,
-				       uint32_t event,
+				       uint64_t event,
 				       struct net_if *iface)
 {
-	LOG_DBG("Event %d", event);
+	LOG_DBG("Event %llu", event);
 	switch (event) {
 		case NET_EVENT_CONN_IF_FATAL_ERROR:
 			LOG_INF("Failed to connect to network");
